--- Packages/com.popo.bdframework/Runtime/AssetsManager/ArtAsset/AssetBundle/V2/AsyncLoadTaskGroup.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 55d2ba4ddbec9fc49a54d0ec80fdedd0
-timeCreated: 1540451298
