+++ Packages/com.popo.bdframework/Runtime/AssetsManager/ArtAsset/AssetBundle/V2/AsyncLoadTaskGroupResult.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 55d2ba4ddbec9fc49a54d0ec80fdedd0
+timeCreated: 1540451298
